Clean code: Move PollingGraphUpdaterParameters out of containing class. This interface is used in many places and is not local the PollingGraphUpdater.
