+++ src/main/java/org/opentripplanner/updater/PollingGraphUpdater.java
@@ -92,22 +92,4 @@ public abstract class PollingGraphUpdater implements GraphUpdater {
         return configRef;
     }
 
-    /**
-     * This is named PollingGraphUpdaterConfig instead of Config in order to not conflict with the
-     * config interfaces of child classes.
-     */
-    public interface PollingGraphUpdaterParameters {
-
-        @Deprecated
-        UpdaterDataSourceParameters getSourceParameters();
-
-        @Deprecated
-        String getUrl();
-
-
-        int getFrequencySec();
-
-        /** The config name/type for the updater. Used to reference the configuration element. */
-        String getConfigRef();
-    }
 }
